iot-subprocess/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-ninja==0.18.0
idna==3.3
pycodestyle==2.8.0
pydantic==1.9.1
requests==2.28.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9